feat: #WB-2569, purge deleted users (wip)
