+++ audience/src/main/java/org/entcore/audience/Audience.java
@@ -1,6 +1,7 @@
 package org.entcore.audience;
 
 import org.entcore.audience.controllers.AudienceController;
+import org.entcore.audience.services.impl.AudienceRepositoryEvents;
 import org.entcore.common.http.BaseServer;
 
 public class Audience extends BaseServer {
@@ -9,5 +10,6 @@ public class Audience extends BaseServer {
   public void start() throws Exception {
     super.start();
     addController(new AudienceController(vertx, config()));
+    setRepositoryEvents(new AudienceRepositoryEvents());
   }
 }
